Resolved the comments left by Bruno and deleted some unnecessary code. Also removed duplicated dependencies. Tested and working as expected
