+++ src/main/java/com/jlr/ttl/ds/api/dto/authentication/JwtConfig.java
@@ -3,6 +3,11 @@ package com.jlr.ttl.ds.api.dto.authentication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * @author PCORREI1
+ *<p>
+ * Configuration class necessary to create a JWT Token Bean. This is needed to run the JWT Token Dependencies.
+ */
 @Configuration
 public class JwtConfig {
