+++ src/main/java/com/jlr/ttl/ds/api/dto/authentication/JwtUtils.java
@@ -1,41 +1,29 @@
 package com.jlr.ttl.ds.api.dto.authentication;
 
-import org.springframework.core.convert.converter.Converter;
-import io.jsonwebtoken.*;
-import org.hibernate.internal.util.collections.ArrayHelper;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
-import java.util.*;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
+/**
+ * @author PCORREI1
+ *<p>
+ * This class allows generating JWT tokens with a specific secret. The tokens are encrypted with a specific SignatureAlgorithm to avoid manual tempering of the token by the user.
+ */
 public class JwtUtils {
+    private final String secret = "secret";
 
-    private String secret = "secret";
-
-    public static class AuthorityAuthenticationProvider implements AuthenticationProvider {
-
-        @Override
-        public Authentication authenticate(Authentication authentication) {
-            return new UsernamePasswordAuthenticationToken(null, null,
-                    authentication.getAuthorities());
-        }
-
-        @Override
-        public boolean supports(Class<?> authentication) {
-            return false;
-        }
-
-    }
-
+    /**
+     * This method is responsible for generating a token with a specific List of permissions.
+     * The claims necessary should be specified in the ListOf argument of the "authorities" claim as new SimpleGrantedAuthority("role").
+     * @return returns the generated token with a list of authorities (endpoints the user can access - to be compared in the Filter) and an expiry date/generation date
+     * **/
     public String generateToken() {
         Map<String, Object> claims = new HashMap<>();
-        claims.put("authorities", List.of(new SimpleGrantedAuthority("Vehicles")));
+        claims.put("authorities", List.of(new SimpleGrantedAuthority("Vehicles"), new SimpleGrantedAuthority("Locations")));
         return Jwts.builder()
                 .setClaims(claims)
                 .setIssuedAt(new Date(System.currentTimeMillis()))
@@ -43,23 +31,4 @@ public class JwtUtils {
                 .signWith(SignatureAlgorithm.HS256, secret)
                 .compact();
     }
-
-    private <T> T getClaimFromToken(String token, Function<Claims, T> claimsResolver) {
-        final Claims claims = getAllClaimsFromToken(token);
-        return claimsResolver.apply(claims);
-    }
-
-    private Claims getAllClaimsFromToken(String token) {
-        return Jwts.parser().setSigningKey(secret).parseClaimsJws(token).getBody();
-    }
-
-//    private Boolean isTokenExpired(String token) {
-//        final Date expiration = getExpirationDateFromToken(token);
-//        return expiration.before(new Date());
-//    }
-
-    private Date getExpirationDateFromToken(String token) {
-        return getClaimFromToken(token, Claims::getExpiration);
-    }
-
 }
