--- src/main/java/com/jlr/ttl/ds/api/dto/entity/JwtResponse.java
@@ -1,19 +1,0 @@
-package com.jlr.ttl.ds.api.dto.entity;
-
-public class JwtResponse {
-
-    private String token;
-
-    public JwtResponse(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-}
